pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Ramil Suleimanov <lllumineux@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.1"
psycopg2 = "^2.9.3"
Pillow = "^9.0.0"
python-dotenv = "^0.19.2"
django-tinymce = "^3.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
